models/yolov8-imagenet/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
ultralytics==8.0.206
torch==2.1.0
torchvision==0.16.0
opencv-python-headless==4.8.1.78
pillow==10.0.1
numpy==1.24.3
requests==2.31.0
pydantic==2.4.2 